Revert into use of spaces for padding freq
This patch addresses the issue detected in: https://github.com/eried/portapack-mayhem/issues/159

This patch will revert the behavior of the function to_string_short_freq

into using spaces on the left of the integer part of the frequency (as it did originally).

When upgrading the scanner app, I did change the behavior of this function eliminating those spaces, so I could gain some characters-worth of space inside the scanner,  but I failed to detect that it introduced some lack of padding on the rx->audio app.

Now, it is back as before, and I also did update the scanner so it can cope with the "extra spaces" this function now adds (again).