Latest-ook-app-with-fixes
This OOK app includes several fixes, including a nasty bug which developer @rascafr noticed, involving the use of auto: on certain type of vars and loops.

The existing (from old HAVOC firmware) debruijn algorithm is also connected into the "DE_BRUIJN TX" button, but I am afraid that the resulting bit stream is NOT debruijn.

Also it does not contemplate the tristate bits (some encoders support 3 values for each address bit: 1, 0 and F ("floating") which in turn multiplies greatly the number of possible combinations.

We are in the need of a better (and working) debruijn algorithm at this time.