Scanner enhancements (issue #986)
The current Scanner app does not allow scrolling through the frequency list using the rotary encoder. I'm changing the PAUSE/RESUME button to allow manual scrolling through frequencies (via the encoder) when focus is on this button. Frequency scrolling will be supported regardless whether or not in PAUSE mode (for forcing a skip to the next frequency before the Wait timer elapses).

The current Wait timer forces a Skip to the next frequency when the timer elapses and a value of 0 means "never lock to a frequency" (which is not useful); I've changed the meaning of "0" to lock on the frequency forever (or until the signal is lost).

Add a second configurable timer setting to maintain lock on a frequency for a specified number of seconds following signal loss. This makes it easier to listen to a conversation on HAM radio (someone will often reply a second or two later) without having to press PAUSE.

I propose to call these two timer fields "Wsa" (wait while signal active) and "Wsl" (wait after signal lost), versus the current "WAIT" field which is more ambiguous.

I propose that the above "Timer" settings will be in 1-second units. (The RECON app uses 1-ms units for two similar timers but I don't think it's necessary nor is there room on the screen)

When a Freq file with a "Range" is read using the "LOAD" button, I propose to populate the Manual Start/End Search Range fields, versus trying to add up to 115 individual frequencies to the Freq List in memory (often overflowing this list meaning the whole range doesn't get searched).

I propose to support Freq files which contain both Range and Single frequencies (but only a single Range in the file due to the change above). Single frequencies will be added to the Freq List, and Range will be loaded into the Start/End Search fields. Note that the current "ADD FQ" button already adds Single frequencies to a file with a Range in it.

When loading from a Freq file, the default Modulation and Step and BW values shall be picked up from the file, if specified, versus leaving the settings on "AM, 9kHz step, 16kHz bw". CAVEAT: At this time I am not (yet) planning to support a separate Modulation type for each frequency in the in-memory Freq List; the first valid value found is used.

Support all BW and Step values in FreqMan (and use the table from FreqMan versus having separate tables in the Scanner code).

Change the "Manual Scan" button to toggle between "SCAN Freq List" & "Sequential Range SEARCH" modes without having to re-LOAD the Freq file. Keep Freq List in memory unless purposely cleared, and allow adding to this in-memory list while performing the Sequential Range Search.

Add a CLEAR button to clear the Freq List in memory without having to load an empty Freq file.

Fixed "FREQ DEL" button to handle removal of frequencies above 4gHz (was incorrectly passing freq in a U32).

Support SQUELCH when scanning is paused, to reduce unnecessary white noise.

Fixed unintentional FWD/REV direction change when switching between scanning Frequency List and performing a Range Search.

Display "REVERSE" or "FORWARD" on the "FWD<>REV" button, indicating which mode will become active if the button is pressed.